requirements-to-freeze.txt
requests
codecov
colorlog
xdg
cathpy
biopython
pyswagger
luigi
